Implement game experience on first load
The first few upgrade costs have been tweaked to guide the player's
attention on the vital first load of the game. There's a new volition rate
upgrade available in the DiscoveryPanel which works differently from
the one in the Improvements tab. This helps game flow. Many variable names
and rendering conditional statements were touched to help direct the
player's attention in stages as more components are unlocked.