Any state variable can now be an upgrade cost
This update took a lot of work. While building, volition was the only
state variable used as an upgrade cost. When implementing an upgrade that
used water, food, and fibers, a render loop crashed the page. Using React
and zustand, if the state variable subscribed to is not a primitive value
like an int or bool, and instead is an object or array, any update to
that value causes a re-render since javascript considers them to have
changed even if the named quantities are the same. Now costs are cached
on the upgradeStore. Keeping the exact object reference allows React and
zustand to function as desired without blowing up in a render loop.