Threats can be encounted while foraging
The "Threat spawner" that worked at random was removed. Now there's a
chance to add a Threat card to the ForagePanel. The upgrade visibility
for the pyramid nav was wired up to the new threat system. The upgrade
to see the pyramid nav is now visible when there's a threat active.