Add SecurityPanel display
The initial view is the "threat status" with an overview of current
threats and protection. An unlock upgrade for Crafting has been added to
the DiscoveryPanel, now visible when the player switches to the Security
tier view. A Crafting view has been added as a toggle button on the
SecurityPanel.